Added a setCell() method and proper imports.
Private grid object can be toggled through the Rullo class. Throws NoSuchElementException when the row or column inputs are out of bounds. Please note that the board toggles between 0 and the space's respective random value. I'll modify the toString() method to output the respective row and column keys so the values can be inputted into the view. I'll also (probably) modify the starting values to all start at 0 instead of the default random value for user-friendliness however, this is a low priority. I still need to build the GUI. Whenever you have a moment to test the new ToggleCell, SumChecker, and Rullo classes to make sure they do as they're supposed to, that would be great. (The exceptions do not need to be tested, they're just there as a sanity check.) I'm going to begin building a a generic grid GUI for Android using the javaFX framework but you should also start building one just in case mine doesn't pan out as intended. I'm going to double check with the teacher to see if the framework NEEDS to be Android, if not, we'll just use javaFX.